+++ www/__tests__/commHelper.test.ts
@@ -1,7 +1,4 @@
-import { mockLogger } from '../__mocks__/globalMocks';
 import { fetchUrlCached } from '../js/services/commHelper';
-
-mockLogger();
 
 // mock for JavaScript 'fetch'
 // we emulate a 100ms delay when i) fetching data and ii) parsing it as text
