+++ www/__tests__/confirmHelper.test.ts
@@ -1,5 +1,3 @@
-import { mockLogger } from '../__mocks__/globalMocks';
-import * as CommHelper from '../js/services/commHelper';
 import {
   baseLabelInputDetails,
   getLabelInputDetails,
@@ -17,7 +15,6 @@ import initializedI18next from '../js/i18nextInit';
 import { CompositeTrip, UserInputEntry } from '../js/types/diaryTypes';
 import { UserInputMap } from '../js/TimelineContext';
 window['i18next'] = initializedI18next;
-mockLogger();
 
 const fakeAppConfig = {
   label_options: 'json/label-options.json.sample',
