+++ www/__tests__/metHelper.test.ts
@@ -1,13 +1,11 @@
 import { getMet } from '../js/metrics/metHelper';
 import { mockBEMUserCache } from '../__mocks__/cordovaMocks';
-import { mockLogger } from '../__mocks__/globalMocks';
 import fakeLabels from '../__mocks__/fakeLabels.json';
 import { getConfig } from '../js/config/dynamicConfig';
 import { initCustomDatasetHelper } from '../js/metrics/customMetricsHelper';
 import fakeConfig from '../__mocks__/fakeConfig.json';
 
 mockBEMUserCache(fakeConfig);
-mockLogger();
 
 global.fetch = (url: string) =>
   new Promise((rs, rj) => {
