+++ www/__tests__/remoteNotifyHandler.test.ts
@@ -1,27 +1,12 @@
 import { EVENTS, publish } from '../js/customEventHandler';
 import { initRemoteNotifyHandler } from '../js/splash/remoteNotifyHandler';
-import {
-  clearURL,
-  getURL,
-  mockBEMUserCache,
-  mockDevice,
-  mockGetAppVersion,
-  mockInAppBrowser,
-} from '../__mocks__/cordovaMocks';
-import { clearAlerts, getAlerts, mockAlert, mockLogger } from '../__mocks__/globalMocks';
-
-mockLogger();
-mockDevice();
-mockBEMUserCache();
-mockGetAppVersion();
-mockInAppBrowser();
-mockAlert();
+import { alerts, clearURL, getURL } from '../__mocks__/cordovaMocks';
 
 const db = window['cordova']?.plugins?.BEMUserCache;
 
 beforeEach(() => {
   clearURL();
-  clearAlerts();
+  alerts.length = 0;
 });
 
 it('does not adds a statEvent if not subscribed', async () => {
@@ -65,12 +50,12 @@ it('handles the popup if subscribed', () => {
       },
     },
   });
-  expect(getAlerts()).toEqual(expect.arrayContaining(['Hello World']));
+  expect(alerts).toEqual(expect.arrayContaining(['Hello World']));
 });
 
 it('does nothing if subscribed and no data', () => {
   initRemoteNotifyHandler();
   publish(EVENTS.CLOUD_NOTIFICATION_EVENT, {});
   expect(getURL()).toEqual('');
-  expect(getAlerts()).toEqual([]);
+  expect(alerts).toEqual([]);
 });
