+++ www/__tests__/startprefs.test.ts
@@ -5,20 +5,6 @@ import {
   getConsentDocument,
 } from '../js/splash/startprefs';
 
-import {
-  mockBEMUserCache,
-  mockBEMDataCollection,
-  mockDevice,
-  mockGetAppVersion,
-} from '../__mocks__/cordovaMocks';
-import { mockLogger } from '../__mocks__/globalMocks';
-
-mockDevice();
-mockGetAppVersion();
-mockBEMUserCache();
-mockBEMDataCollection();
-mockLogger();
-
 global.fetch = (url: string) =>
   new Promise((rs, rj) => {
     setTimeout(() =>
